refactor: move FindPatientCommand into command package
