+++ src/test/java/seedu/address/logic/parser/patient/FindPatientCommandParserTest.java
@@ -1,16 +1,15 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.patient;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
-import static seedu.address.logic.parser.patient.FindPatientCommand.MESSAGE_USAGE;
+import static seedu.address.logic.commands.patient.FindPatientCommand.MESSAGE_USAGE;
 
 import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.parser.patient.FindPatientCommand;
-import seedu.address.logic.parser.patient.FindPatientCommandParser;
+import seedu.address.logic.commands.patient.FindPatientCommand;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 
 public class FindPatientCommandParserTest {
